Collapse bounds calculations into DisplayListBuilder
